+++ src/NSimulate.UnitTest/Instruction/DeactivateInstructionFixture.cs
@@ -19,7 +19,7 @@ namespace NSimulate.UnitTest
 
 				var instruction = new DeactivateInstruction(testProcess);
 
-				int? nextTimePeriodCheck = null;
+				long? nextTimePeriodCheck = null;
 				bool canComplete = instruction.CanComplete(context, out nextTimePeriodCheck);
 
 				Assert.IsTrue(canComplete);
