+++ src/NSimulate.UnitTest/Instruction/InterruptInstructionFixture.cs
@@ -25,7 +25,7 @@ namespace NSimulate.UnitTest
 
 				var instruction = new InterruptInstruction(testProcess);
 
-				int? nextTimePeriodCheck = null;
+				long? nextTimePeriodCheck = null;
 				bool canComplete = instruction.CanComplete(context, out nextTimePeriodCheck);
 
 				Assert.IsTrue(canComplete);
